put and delete routes for webhooks

+++ src/components/Domain/index.ts
@@ -121,6 +121,10 @@ export async function remove(
     next: NextFunction
   ): Promise<void> {
     try {
+
+        const user: IUserModel = await AuthService.authUser(req);
+        if (!user) throw new Error('unauthorized user'); 
+
         await DomainService.remove(req.params.id);
         res.status(200).json({success: true });
